app-history/navigate/goTo-repeated.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(async () => {
    let key = appHistory.current.key;
    await appHistory.navigate("#1");

    let first_promise_resolved = false;
    let second_promise_resolved = false;
    let first_promise = appHistory.goTo(key).then(() => first_promise_resolved = true);
    let second_promise = appHistory.goTo(key).then(() => second_promise_resolved = true);
    assert_false(first_promise_resolved);
    assert_false(second_promise_resolved);
    await first_promise;
    t.step_func(() => {
      assert_true(first_promise_resolved);
      assert_true(second_promise_resolved);
    })();
    t.done();
  }, 0);
}, "Repeated navigate.goTo()");
</script>
